move analytics methods to a concern
