+++ doc/gradable.md
@@ -18,8 +18,8 @@ When an instance of a model that `include`s Gradable is destroyed, its grades an
 *Note that when discussing grades in an assignment, we are always referring to grades that have either been graded or released*
 
   * `no_grade(attrs)` - used for rejecting the nested attributes of a grade. When the model's `pass_fail` attribute is true, the method returns true if the grade's pass fail status is empty. When `pass_fail` is false, the method returns true if the grade's `raw_points` attribute has not been set yet
-  * `graded_or_released_scores` - returns an array consisting of the `raw_points` values from every grade in the assignment
-  * `grade_count` - number of grades in the assignment
+  * `graded_or_released_scores` - returns an array consisting of the `raw_points` values from every grade in the assignment (moved to `Scorable` concern)
+  * `grade_count` - number of grades in the assignment (moved to `Scorable` concern)
   * `grade_for_student(student)` - returns the passed in student's grade for the assignment. Returns nil if the grade has neither been graded nor released, or if the grade does not yet exist
   * `average` - returns the average number of raw points out of all grades belonging to the assignment
   * `earned_average` - returns the average score of all grades in the assignment that have scores greater than zero
